+++ controller/jogoController.py
@@ -1,31 +1,73 @@
 from connectStore import app
 from repository.jogoRepository import *
 from flask import jsonify, request
+from utils import validaToken
 
 @app.route('/jogo/', methods=['GET'])
 def getAllJogos():
-    jogos = findAllJogo()
+    data = request.headers.get('Authorization')
 
-    return jsonify(jogos)
+    if data:
+
+        if validaToken(data):
+            return jsonify(findAllJogo())
+
+        return None
+
+    return 'Acesso nao autorizado'
+
+
 
 @app.route('/jogo/<int:id>', methods=['GET'])
 def getByIdJogos(id: int):
-    jogo = findByIdJogo(id)
+    data = request.headers.get('Authorization')
 
-    return jsonify(jogo)
+    if data:
+
+        if validaToken(data):
+            return jsonify(findByIdJogo(id))
+
+        return None
+
+    return 'Acesso nao autorizado'
 
 @app.route('/jogo/criar-jogo', methods=['POST'])
 def criarJogos():
     data = request.get_json(silent=True)
+    token = request.headers.get('Authorization')
 
-    return postJogo(data)
+    if token:
+
+        if validaToken(token):
+            return postJogo(data)
+
+        return None
+
+    return 'Acesso nao autorizado'
 
 @app.route('/jogo/atualizar-jogo', methods=['PUT'])
 def atualizarJogos():
     data = request.get_json(silent=True)
+    token = request.headers.get('Authorization')
 
-    return putJogo(data)
+    if token:
+
+        if validaToken(token):
+            return putJogo(data)
+
+        return None
+
+    return 'Acesso nao autorizado'
 
 @app.route('/jogo/deletar-jogo/<int:id>', methods=['DELETE'])
 def deletarJogos(id):
-    return deleteByIdJogo(id)
+    token = request.headers.get('Authorization')
+
+    if token:
+
+        if validaToken(token):
+            return deleteByIdJogo(id)
+
+        return None
+
+    return 'Acesso nao autorizado'
